+++ src/c.h
@@ -1,5 +1,0 @@
-#include <errno.h>
-#include <lua.h>
-#include <lauxlib.h>
-#include <lualib.h>
-#include <linenoise.h>